requirements.txt
streamlit
smartapi-python
pandas
ta
backtrader
fastapi
uvicorn
python-telegram-bot
finrl
pyotp
logzero
